requirements.txt
asgiref==3.4.1
autopep8==1.5.7
bcrypt==3.2.0
cffi==1.14.6
click==8.0.1
colorama==0.4.4
cryptography==3.4.7
ecdsa==0.17.0
fastapi==0.68.0
greenlet==1.1.1
h11==0.12.0
httptools==0.2.0
passlib==1.7.4
pyasn1==0.4.8
pycodestyle==2.7.0
pycparser==2.20
pydantic==1.8.2
python-dotenv==0.19.0
python-jose==3.3.0
PyYAML==5.4.1
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.4.22
starlette==0.14.2
toml==0.10.2
typing-extensions==3.10.0.0
uvicorn==0.14.0
watchgod==0.7
websockets==9.1
